New release workflow
Add a GH Actions workflow that runs when a GitHub release is published.
It will run the new invoke task "update-version" with the tag name,
updating `metadata.json`'s `"version"` value with the correct version.
Then it will update the tag to the new commit and push it all to
`master`.